Extract the LaF Swing preview into a separate class
This introduces a clear distinction between the Swing and the SWT part
and thus reduces the chance of accessing the Swing component outside of
the AWT event thread.
